requirements.txt
yfinance
opendartreader
finance-datareader==0.9.96

mcp[cli]
httpx
toolbox-core

matplotlib
seaborn

google-generativeai